+++ app/models/mattermost/bot_config.rb
@@ -4,8 +4,44 @@ module Mattermost
   module BotConfig
     module_function
 
+    # Plugin id is "openproject-mattermost". OpenProject stores that as
+    # plugin_openproject-mattermost (hyphen) and/or plugin_openproject_mattermost.
+    SETTING_KEYS = [
+      "plugin_openproject_mattermost",
+      "plugin_openproject-mattermost"
+    ].freeze
+
     def settings
-      Hash(Setting.plugin_openproject_mattermost).with_indifferent_access
+      Hash(raw_settings).with_indifferent_access
+    end
+
+    def raw_settings
+      SETTING_KEYS.each do |key|
+        val = begin
+          Setting[key]
+        rescue StandardError
+          nil
+        end
+        return val if val.present?
+      end
+      if Setting.respond_to?(:plugin_openproject_mattermost)
+        val = Setting.plugin_openproject_mattermost
+        return val if val.present?
+      end
+      {}
+    end
+
+    def write!(hash)
+      payload = Hash(hash).stringify_keys
+      SETTING_KEYS.each do |key|
+        Setting[key] = payload
+      rescue StandardError
+        nil
+      end
+      if Setting.respond_to?(:plugin_openproject_mattermost=)
+        Setting.plugin_openproject_mattermost = payload
+      end
+      payload
     end
 
     def server_url
@@ -22,6 +58,17 @@ module Mattermost
 
     def ready?
       server_url.present? && bot_token.present?
+    end
+
+    def debug_keys
+      SETTING_KEYS.map do |key|
+        val = begin
+          Setting[key]
+        rescue StandardError
+          nil
+        end
+        "#{key}=#{val.present? ? 'set' : 'empty'}"
+      end.join(",")
     end
 
     def client